# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Earscope API",
        "version": "1.0"
    },
    "paths": {
...
        "/v1/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get all videos",
                "operationId": "2dd39694f233ae275064d4e27e73aea0",
                "responses": {
                    "200": {
                        "description": "List of all videos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "raw_video_url": {
                                                "type": "string"
                                            },
                                            "processed_video_url": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Upload a pair of videos (raw and processed)",
                "operationId": "766f06dc2935b23771613a91839bb1a1",
                "requestBody": {
                    "description": "Upload raw and processed videos",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "raw_video",
                                    "processed_video"
                                ],
                                "properties": {
                                    "raw_video": {
                                        "description": "Raw video file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "processed_video": {
                                        "description": "Processed video file with bounding box",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Videos uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "raw_video_path": {
                                                    "type": "string"
                                                },
                                                "processed_video_path": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/{videoId}/assign/{userId}": {
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Assign a video to a user",
                "operationId": "ea5bac4dbe008179d0e71352b804c1ef",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video assigned successfully"
                    },
                    "404": {
                        "description": "Video or User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/{videoId}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get video details by ID",
                "operationId": "eeed8b39f6ddcc2f00bfaff8dec07955",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "raw_video_url": {
                                            "type": "string"
                                        },
                                        "processed_video_url": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Videos"
                ],
                "summary": "Unassign a video",
                "operationId": "3ab86baf997ffc3d718acfe4a42cce06",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video unassigned successfully"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/pasien": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get videos assigned to the logged-in patient",
                "operationId": "1f76df1bad6c346cd8a2e2831303dea0",
                "responses": {
                    "200": {
                        "description": "List of videos assigned to the patient",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "raw_video_url": {
                                                "type": "string"
                                            },
                                            "processed_video_url": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/stream/{filename}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Stream a video file",
                "operationId": "1b927ddc0005c5ebec55da753647ef4a",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "The name of the video file to stream",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video streaming successfully",
                        "headers": {
                            "Content-Type": {
                                "description": "MIME type of the video",
                                "schema": {
                                    "type": "string",
                                    "example": "video/mp4"
                                }
                            },
                            "Accept-Ranges": {
                                "description": "bytes range support",
                                "schema": {
                                    "type": "string",
                                    "example": "bytes"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token di kolom ini dengan format Bearer <token>"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}